Refactor shared utility classes
