+++ apps/data-streams/src/app.module.ts
@@ -2,10 +2,10 @@ import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { Config } from './app.config.service';
 import { WorkerConnectionFactory } from './app.worker.connection.factory';
 import { LoggerModule } from 'nestjs-pino';
 import { TcpController } from './app.tcp.controller';
+import { Config } from 'apps/util/config.service';
 
 @Module({
   imports: [
