+++ apps/data-streams/src/app.worker.connection.factory.ts
@@ -1,7 +1,7 @@
 import { Logger } from "@nestjs/common";
 import { ClientProxyFactory, Transport } from "@nestjs/microservices";
 import { LoggerFactory } from "apps/util/util.logger.factory";
-import { Config } from "./app.config.service";
+import { Config } from "../../util/config.service";
 
 export class WorkerConnectionFactory {
