+++ apps/util/config.service.ts
@@ -4,6 +4,7 @@ import { ConfigService } from '@nestjs/config';
 @Injectable()
 export class Config {
     readonly NODE_ENV: string;
+    readonly APP_HOST: string;
     readonly APP_HTTP_PORT: number;
     readonly APP_TCP_PORT: number;
     readonly WORKER_HOST: string;
@@ -11,6 +12,7 @@ export class Config {
 
     constructor(@Inject(ConfigService) private configService: ConfigService) {
         this.NODE_ENV = configService.get<string>('NODE_ENV')
+        this.APP_HOST = configService.get<string>('APP_HOST')
         this.APP_HTTP_PORT = parseInt(configService.get<string>('APP_HTTP_PORT'))
         this.APP_TCP_PORT = parseInt(configService.get<string>('APP_TCP_PORT'))
         this.WORKER_HOST = configService.get<string>('WORKER_HOST')
