+++ apps/worker/src/worker.config.dto.ts
@@ -1,4 +1,4 @@
-import { IsDefined, IsIn, IsNotEmpty, IsNumber, IsNumberString, IsOptional, IsPositive, Min } from "class-validator";
+import { IsDefined, IsIn, IsNotEmpty, IsNumber, IsPositive, Min } from "class-validator";
 
 export class WorkerConfigDto {
     @IsIn(['IQAIR_DAILY'])
